Aux branch

+++ src/routes/admin/sizes/+page.server.ts
@@ -1,12 +1,16 @@
 import type { Actions, PageServerLoad } from "./$types";
 import type { Size } from "$lib/server/types/models";
-import { fail } from "@sveltejs/kit";
+import { error, fail } from "@sveltejs/kit";
 
 
 export const load: PageServerLoad = async({ fetch }) => {
     const response = await fetch(`/api/size`);
-    
+
     const sizes: Size[] = await response.json();
+
+    if (!response.ok) {
+        return error(404, 'Failed to fetch sizes');
+    }    
 
     return { sizes }
 };
@@ -16,6 +20,8 @@ export const actions = {
     default: async ({ request, fetch }) => {
         const formData = await request.formData();
         const id = formData.get('id');
+
+        if (!id || typeof id !== 'string') return fail(400, { error: 'Invalid size ID' });
 
         const response = await fetch(`/api/size/${id}`, {
             method: 'DELETE',
